sky map draw

+++ app/src/main/java/com/vaonis/skymap/businesslogic/AstronomicalObject.kt
@@ -2,7 +2,7 @@ package com.vaonis.skymap.businesslogic
 
 data class AstronomicalObject(
     val id: String,
-    val ra: Double, val de: Double,
+    val ra: Float, val de: Float,
     val category: String,
     val distance: Long,
     val distanceUnit: DistanceUnit
